+++ packages/mobx-classic/api/extendobservable.ts
