+++ packages/mobx-classic/api/intercept.ts
