+++ packages/mobx-classic/api/iscomputed.ts
