+++ packages/mobx-classic/api/isobservable.ts
