+++ packages/mobx-classic/tsconfig.json
@@ -1,0 +1,26 @@
+{
+    "version": "2.0.10",
+    "compilerOptions": {
+        "target": "es5",
+        "outDir": "lib",
+        "sourceMap": false,
+        "declaration": true,
+        "module": "commonjs",
+        "removeComments": false,
+        "strictNullChecks": true,
+        "noImplicitReturns": true,
+        "noImplicitUseStrict": true,
+        "noImplicitAny": false,
+        "moduleResolution": "node",
+        "lib": [
+            "es5",
+            "dom"
+        ]
+    },
+    "exclude": [
+        "node_modules",
+        "test",
+        "lib",
+        ".build"
+    ]
+}
